refactor: mv time series dataset class to utils
